% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sim.R
\name{data_sim}
\alias{data_sim}
\title{data simulation for ERF example Data}
\usage{
data_sim(
  ncell = 100,
  ncov = 5,
  prob_missing = 0.95,
  unif.bnd = c(10, 30),
  mat.var = 0.05,
  nsamp = 10000
)
}
\arguments{
\item{ncell}{number of spatial grid cells}

\item{ncov}{number of covariates to simulate}

\item{prob_missing}{probability of nondetection (1-detection)}

\item{mat.var}{variance of the Matérn autocorrelation}

\item{nsamp}{number of random samples to draw}

\item{unif.bnds}{bounds of the uniform distribution draws of the scale parameter for the Matérn autocorrelation}
}
\value{
A list containing the samples, the beta coefficients, and the raster brick
}
\description{
A function to calculate various ROC curve performance metrics given model predictions and the oberved samples
}
\examples{
library(RandomFields)
library(raster)
df <- data_sim()
head(df$samples)
table(df$samples$obs)

colr <- colorRampPalette(c('dodgerblue4','dodgerblue2','ivory','firebrick2','firebrick4'))
plot(df$grid$prob, col=colr(100), xaxt='n', yaxt='n', asp=NA)
with(df$samples[df$samples$obs==1,], points(x,y,pch=16))

}
