% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_splitter.R
\name{max_splitter}
\alias{max_splitter}
\title{Determines the maximum sample size in the second RF bag}
\usage{
max_splitter(v, p = 0.89, nmax = 10000)
}
\arguments{
\item{v}{A data frame object returned from erf_data_prep()}

\item{p}{A numeric value between (0,1), default is just below 90\%}

\item{nmax}{A numeric value > 0 specifying the maximum number of observations per bag, default is 1e4}
}
\value{
A numeric value specifying the maximum split
}
\description{
Takes a data.frame object determines the maximum split
}
\examples{
data <- erf_data_prep(
				df=simData$samples, 
				var='obs', 
				covariates=grep('cov', 
					colnames(simData$samples), value=TRUE)
			)
max_splitter(data)
max_splitter(data, p=0.6)

}
