% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_splitter.R
\name{min_splitter}
\alias{min_splitter}
\title{Determines the minimum sample size in the first RF bag}
\usage{
min_splitter(v, p = 0.89, nmax = 10000)
}
\arguments{
\item{v}{A data frame object returned from erf_data_prep()}

\item{p}{A numeric value between (0,1), default is just below 90\%}

\item{nmax}{A numeric value > 0 specifying the maximum number of observations per bag, default is 1e4}
}
\value{
A numeric value specifying the minimum split
}
\description{
Takes a data.frame object determines the minimum split
}
\examples{
v <- data.frame(var=rbinom(1000,1,0.2), covar=rnorm(1000))
min_split(v)

}
