% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ALE.R
\name{plot_ALE}
\alias{plot_ALE}
\title{Plots Accumulated Local Effects for ERF}
\usage{
plot_ALE(ALE, name, cex.axis = 1, cex.lab = 1, quantiles = c(0.025, 0.975))
}
\arguments{
\item{ALE}{an ALE object for a given variable}

\item{name}{the name of the variable}

\item{cex.axis}{the cex of the axis}

\item{cex.lab}{the cex of the labels}
}
\value{
A plot of the ALEs
}
\description{
Plots the Accumulated Local Effects (ALE) from an ERF object
}
\examples{
#run an ERF with 10 RFs and 
ens_rf_ex <- ens_random_forests(df=simData$samples, var="obs", covariates=grep("cov", colnames(simData$samples),value=T), save=FALSE, cores=1)

ALEdf <- ALE_fn(ens_rf_ex, save=FALSE)

plot_ALE(ALEdf[1])

}
