% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_ens_fn.R
\name{rf_ens_fn}
\alias{rf_ens_fn}
\title{Random Forests model with additional bagging}
\usage{
rf_ens_fn(v, var, form, min_split, ntree = 1000, mtry = 5, importance = TRUE)
}
\arguments{
\item{v}{A data frame object created by \code{erf_data_prep()} or internally in \code{erf()}}

\item{var}{A character string specifying the column name to use as the dependent variable}

\item{form}{A formula class object specifying the RF model formulation (created by \code{erf_formula_prep()} or internal in \code{erf()})}

\item{min_split}{The minimum number of samples in the RF bagging procedure (created internally by \code{erf()})}

\item{ntree}{The number of decision trees to use in each RF, default is 1000}

\item{mtry}{The number of covariates to try at each node split, default is 5}

\item{importance}{A logical flag for the randomForest model to calculate the variable importance}
}
\value{
A list containing mod (the randomForest model), preds (the predictions), roc_train (the Receiver Operator Characteristic Curve performance metrics calculated by rocr_ens() on the training set), roc_test (the Receiver Operator Characteristic Curve performance metrics calculated by rocr_ens() on the test set)
}
\description{
Runs a single Random Forests model with an additional bagging layer
}
\examples{

}
