% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocr_ens.R
\name{rocr_ens}
\alias{rocr_ens}
\title{Receiver Operator Characteristic Curve calculations}
\usage{
rocr_ens(predictions, samples)
}
\arguments{
\item{predictions}{Random Forest predictions (or any probability)}

\item{samples}{The original observations (must be a factor)}
}
\value{
A list containing all the various ROC curve performance metrics
}
\description{
A function to calculate various ROC curve performance metrics given model predictions and the oberved samples
}
\examples{
predictions <- rbeta(100, 4, 4)
samples <- rbinom(100, 1, 0.5)
rocr <- rocr_ens(predictions, samples)
str(rocr)
rocr$auc

}
