% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALE_fn.R
\name{ALE_fn}
\alias{ALE_fn}
\title{Accumulated Local Effects for ERF}
\usage{
ALE_fn(
  fit,
  var,
  save = TRUE,
  out.folder = NULL,
  cores = parallel::detectCores() - 4
)
}
\arguments{
\item{fit}{The fitted object returned from calling ens_random_forests()}

\item{var}{The name of the response variable}

\item{save}{A logical flag to save the output as an RData object, default is TRUE.}

\item{out.folder}{A path to the folder to write out too. If NULL then a folder is generated in the working directory}

\item{cores}{An integer value that either indicates the number of cores to use for parallel processing or a negative value to indicate the number of cores to leave free. Default is to leave two cores free.}
}
\value{
A list that contains a data.frame for each variable, ordered by the mean variable importance, and a vector of the covariate values (used for rug plot in plot_ALE). The columns in each data.frame are as follows:
\itemize{
\item \strong{x}: the covariate values that the ALE was calculated for
\item \strong{class}: the class of the covariate; used by subsequent plot_ALE function
\item \strong{q}: the quantile of the x value of the covariate
\item \strong{f.X}: the ALEs evaluated at a given x value
}
}
\description{
Calculates the Accumulated Local Effects (ALE) from an ERF object
}
\examples{
#run an ERF with 10 RFs and 
ens_rf_ex <- ens_random_forests(df=simData$samples, var="obs", covariates=grep("cov", colnames(simData$samples),value=T), save=FALSE, cores=1)

ALEdf <- ALE_fn(ens_rf_ex, save=FALSE)
head(ALEdf[[1]]$df)

}
