% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALE_fn.R
\name{ALE_fn}
\alias{ALE_fn}
\title{Accumulated Local Effects for ERF}
\usage{
ALE_fn(
  fit,
  var,
  save = TRUE,
  out.folder = NULL,
  cores = parallel::detectCores() - 4,
  quants = c(0.05, 0.95)
)
}
\arguments{
\item{fit}{The fitted object returned from calling ens_random_forests()}

\item{var}{The name of the response variable}

\item{save}{A logical flag to save the output as an RData object, default is TRUE.}

\item{out.folder}{A path to the folder to write out too. If NULL then a folder is generated in the working directory}

\item{cores}{An integer value that either indicates the number of cores to use for parallel processing or a negative value to indicate the number of cores to leave free. Default is to leave two cores free.}

\item{quants}{Either a NULL or two numeric values specifying the lower and upper quantile cutoffs to evaluate the ALE for continuous covariates over.}
}
\value{
A list that contains the ALEs for each variable, ordered by the mean variable importance
}
\description{
Calculates the Accumulated Local Effects (ALE) from an ERF object
}
\examples{
#run an ERF with 10 RFs and 
ens_rf_ex <- ens_random_forests(df=simData$samples, var="obs", covariates=grep("cov", colnames(simData$samples),value=T), save=FALSE, cores=1)

ALEdf <- ALE_fn(ens_rf_ex, save=FALSE)
head(ALEdf[[1]])

}
