% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_random_forests.R
\name{ens_random_forests}
\alias{ens_random_forests}
\title{Ensemble Random Forests}
\usage{
ens_random_forests(
  df,
  var,
  covariates,
  out.folder = NULL,
  duplicate = TRUE,
  n.forests = 10,
  cores = parallel::detectCores() - 2,
  save = TRUE
)
}
\arguments{
\item{df}{A data.frame object}

\item{var}{A character string indicating the column name of the data frame that contains the number of interactions for the ERF to model; column should be a numeric column}

\item{covariates}{A character vector indicating the column name(s) of the data frame that contain the covariates}

\item{out.folder}{A path to the folder to write out too. If NULL then a folder is generated in the working directory}

\item{duplicate}{A logical flag that indicates whether to duplicate observations with more than one interaction. Default is TRUE to duplicate all records that interacted with more than one individual (i.e. a fishing set that caught two of the same species)}

\item{n.forests}{A numeric value indicating how many Random Forests to generate in the ensemble, default is 100}

\item{cores}{A numeric value that either indicates the number of cores to use for parallel processing or a negative value to indicate the number of cores to leave free. Default is to leave two cores free.}

\item{save}{A logical flag to save the output as an RData object, default is TRUE.}
}
\value{
A fitted ERF model as a list object. Contains the model data, the fitted RF models, the ensemble predictions, the ensemble training/test performance, the ROC training/test performance for each RF, and the individual RF predictions.
}
\description{
Execute Ensemble Random Forests model on a given dataset
}
