% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_ens_fn.R
\name{rf_ens_fn}
\alias{rf_ens_fn}
\title{Random Forest model with additional bagging}
\usage{
rf_ens_fn(v, form, max_split, ntree = 100, mtry = 5, importance = TRUE)
}
\arguments{
\item{v}{A data frame object created by \code{erf_data_prep()} or internally in \code{ens_random_forests()}}

\item{form}{A formula class object specifying the RF model formulation (created by \code{erf_formula_prep()} or internal in \code{ens_random_forests()})}

\item{max_split}{The maximum number of samples in the RF bagging procedure (created internally by \code{ens_random_forests()})}

\item{ntree}{The number of decision trees to use in each RF, default is 100}

\item{mtry}{The number of covariates to try at each node split, default is 5}

\item{importance}{A logical flag for the randomForest model to calculate the variable importance}
}
\value{
A list containing mod (the randomForest model), preds (the predictions), roc_train (the Receiver Operator Characteristic Curve performance metrics calculated by rocr_ens() on the training set), roc_test (the Receiver Operator Characteristic Curve performance metrics calculated by rocr_ens() on the test set)
}
\description{
Runs a single Random Forest model with an additional bagging layer and calculates performance metrics
}
\examples{
form <- erf_formula_prep('obs', grep('cov',colnames(simData$samples),value=TRUE))
data <- erf_data_prep(simData$samples, 'obs', grep('cov', colnames(simData$samples), value=TRUE))
max_split <- max_splitter(data)

#fit a single RandomForest
rf_ex <- rf_ens_fn(data, form, max_split, ntree=50)

#see the training/test auc value
rf_ex$roc_train$auc
rf_ex$roc_test$auc

#see the distribution of predictions
par(mar=c(4,4,1,1))
plot(density(rf_ex$preds[,2],from=0,to=1,adj=2), main="")

}
